requirements.txt
Django==5.2.10
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
PyJWT==2.10.1
python-dotenv==1.2.1
PyMySQL==1.1.2
gunicorn==23.0.0
